Move the last package out of certification
Only one package remained in certification. There can be only none.

* Moves artifacts to internal
