+++ internal/artifacts/filesystem_writer.go
