+++ internal/artifacts/filesystem_writer_test.go
