+++ internal/artifacts/map_writer.go
