+++ internal/artifacts/map_writer_test.go
