+++ internal/lib/logging_test.go
@@ -8,7 +8,7 @@ import (
 	. "github.com/onsi/gomega"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/artifacts"
+	"github.com/redhat-openshift-ecosystem/openshift-preflight/internal/artifacts"
 )
 
 var _ = Describe("Lib Logging Functions", func() {
