+++ internal/operatorsdk/operatorsdk_test.go
@@ -7,7 +7,7 @@ import (
 	"os/exec"
 	"testing"
 
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/artifacts"
+	"github.com/redhat-openshift-ecosystem/openshift-preflight/internal/artifacts"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
