+++ internal/policy/operator/deployable_by_olm_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification"
-	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/artifacts"
+	"github.com/redhat-openshift-ecosystem/openshift-preflight/internal/artifacts"
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/internal/openshift"
 
 	fakecranev1 "github.com/google/go-containerregistry/pkg/v1/fake"
